(j9jitbuilder) Add JitBuilderMethod resolved method subclass
In preparation for being able to build OMRIlBuilder.cpp in OpenJ9,
the JitBuilder notions of ResolvedMethod need to be established.
First step is to subclass TR_ResolvedJ9Method and bring in
the JitBuilder notion of a resolved method (now called
J9::JitBuilderMethod). To facilitate contruction, some new
protected constructors were introduced in base classes
TR_ResolvedJ9MethodBase and TR_J9Method. Second step is
to create TR::ResolvedMethod (which has not been previously
used in OpenJ9) as a subclass of J9::JitBuilderMethod.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>