(j9jitbuilder) Prepare front end and memory classes
OMRIlBuilder expects a TR::FrontEnd class to exist which
needs to be a concrete object. OpenJ9 uses the unrelated
TR_FrontEnd base class with multiple subclasses. To get
TR::FrontEnd, I added the env/ConcreteFE.hpp class which
pulls in the front end classes and then creates TR::FrontEnd
as a subclass of TR_J9VM with the same set of contructors.
To enable OMRIlBuilder to construct its objects, a new
constructor API is needed for the OpenJ9 SystemSegmentProvider
class (with fewer arguments).  Define TR::SystemSegmentProvider
class adding a constructor that automatically creates a
J9::SegmentAllocator, pre-configures the segment sizes, and
currently sets no allocation limit. Some mechanism added in
this class to make sure that the J9::SegmentAllocator is
properly destroyed and freed when this particular
constructor is used. This change will become unnecessary once
the OMR and OpenJ9 memory allocation story is unified.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>