(j9jitbuilder) Add support to add and handle MethodBuilder compiles
Adds _methodBuilder to the MethodEvent type and some code to
detect if a method queue entry is a MethodBuilder compilation
request. Creates the MethodBuilderDetails object and requests
the compilation using that object. Adds code to handle this
event to create the optimization plan. Finally, short-circuits
many code paths that assume the method being compiled has a
J9Method or an underlying J9Class that can be queried. Probably
a better long term solution will be to add MethodBuilder methods
to an artificial class with artificial method, but for now it's
easier to just avoid dereferencing the compilation's J9Method /
J9Class.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>