Test JitBuilder compilation inside OpenJ9 JIT
Creates a simple method dynamically that returns the
uninteresting 64-bit integer "3" as the simplest
JitBuilder based method you can compile and the standard
"first" attempt.

The MethodBuilder is created, compiled, executed, and
result tested in jitHookAboutToRunMain (so it will only
happen if run with -Xjit:noJitUntilMain).  Also adds
initializeJit and shutdownJit calls in the appropriate
hooks in rossa.cpp. These are needed to initialize the
JitBuilder client API allocators. The internal
initializeJit and shutdownJit calls are empty so they do
not actually perform any initialization or shutdown of
the OpenJ9 JIT.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>