Insert new synchronized jitGetExceptionTableFromPC path as fallback
jitGetExceptionTableFromPC is not synchronized, this can cause a race-
condition loading exceptionTable/jitInfo if the JIT is updating it.

The race-condition is rare and synchronizing jitGetExceptionTableFromPC
can be costly, hence this implements a synchronized fallback path.

New structure:
- Stack-walker loads exceptionTable/jitInfo (non-synchronized)
- If NULL it assumes it's an interpreter transition
- Checks PC in i2jReturnTable
- If not in i2jReturnTable:
  - [NEW] Use synchronized loading of exceptionTable/jitInfo
  - Report invalidJITReturnAddress error

Implementation uses a new field in J9JavaVM struct to pass the
JIT-ArtifactMonitor to the stack-walker.